+++ systemimage/config.py
@@ -48,6 +48,10 @@ def port_value_converter(value):
     return result
 
 
+def device_converter(value):
+    return value.strip()
+
+
 class Configuration:
     def __init__(self, ini_file=None):
         # Defaults.
@@ -82,7 +86,9 @@ class Configuration:
         self.config_file = path
         self.service.update(converters=dict(http_port=port_value_converter,
                                             https_port=port_value_converter,
-                                            build_number=int),
+                                            build_number=int,
+                                            device=device_converter,
+                                            ),
                            **parser['service'])
         if (self.service.http_port is DISABLED and
             self.service.https_port is DISABLED):
@@ -170,9 +176,15 @@ class Configuration:
 
     @property
     def device(self):
-        # It's safe to cache this.
         if self._device is None:
-            self._device = self.hooks.device().get_device()
+            # Start by looking for a [service]device setting.  Use this if it
+            # exists, otherwise fall back to calling the hook.
+            self._device = getattr(self.service, 'device', None)
+            # The key could exist in the channel.ini file, but its value could
+            # be empty.  That's semantically equivalent to a missing
+            # [service]device setting.
+            if not self._device:
+                self._device = self.hooks.device().get_device()
         return self._device
 
     @device.setter